Migrate to playwright test runner
