+++ package.json
@@ -1,6 +1,7 @@
 {
   "devDependencies": {
     "@parcel/transformer-sass": "2.0.0-beta.2",
+    "@playwright/test": "^1.12.2",
     "@types/mime-types": "^2.1.0",
     "mime-types": "^2.1.31",
     "playwright": "^1.12.2",
