+++ playwright.config.ts
@@ -1,0 +1,25 @@
+import { PlaywrightTestConfig } from '@playwright/test';
+const config: PlaywrightTestConfig = {
+    use: {
+        headless: true,
+        viewport: { width: 1280, height: 720 },
+        ignoreHTTPSErrors: true,
+    },
+    projects: [
+        {
+            name: 'Chromium',
+            use: {
+                browserName: 'chromium',
+            },
+        },
+        {
+            name: 'Firefox',
+            use: { browserName: 'firefox' },
+        },
+        {
+            name: 'WebKit',
+            use: { browserName: 'webkit' },
+        },
+    ],
+};
+export default config;
